Extract CtaFooter to client component, make landing page a Server Component
Addresses review findings:
- Only CtaFooter needs "use client" (it uses useState for the email form)
- The rest of the page (Nav, Hero, Features, Architecture, Stats) is now
  server-rendered, reducing the client JS bundle significantly
- Fix absolute positioning bug on footer glow overlay (missing relative parent)
- Add noopener to external link rel attribute
- Use stable keys for terminal lines instead of array index
- Add TODO comment on stub email form
